requirements-prod.txt
torch==2.5.1
torchaudio==2.5.1
zonos @ git+https://github.com/UpperMoon0/nstut-zonos-fork.git@main
tokenizers==0.21.0
transformers>=4.30.0
huggingface_hub>=0.17.0
websockets==11.0.3
python-dotenv==1.0.0
numpy==2.2.2
soundfile==0.13.1
aiohttp>=3.10.5
# replace with triton-windows if you are on Windows
triton
edge-tts>=6.1.10
fastapi
uvicorn[standard]
pydantic
psutil